pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "youtube-thumbnail-generator"
version = "2.6.2"
authors = [
    {name = "preangelleo"}
]
description = "A powerful YouTube thumbnail generator with AI text optimization and multi-language support"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.9"
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "Topic :: Multimedia :: Graphics",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Operating System :: OS Independent",
]
keywords = ["youtube", "thumbnail", "generator", "ai", "gemini", "image", "processing"]
dependencies = [
    "Pillow>=10.0.0",
    "numpy>=1.24.0",
    "python-dotenv>=1.0.0",
    "Flask>=2.3.0",
    "flask-cors>=4.0.0",
    "google-generativeai>=0.3.0",
    "langdetect>=1.0.9",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-cov>=4.1.0",
    "black>=23.7.0",
    "flake8>=6.1.0",
    "mypy>=1.5.0",
]

[project.urls]
"Homepage" = "https://github.com/preangelleo/youtube-thumbnail-generator"
"Bug Reports" = "https://github.com/preangelleo/youtube-thumbnail-generator/issues"
"Source" = "https://github.com/preangelleo/youtube-thumbnail-generator"
"Documentation" = "https://github.com/preangelleo/youtube-thumbnail-generator/docs"

[project.scripts]
youtube-thumbnail = "youtube_thumbnail_generator.cli:main"
youtube-thumbnail-api = "youtube_thumbnail_generator.api:main"

[tool.setuptools.packages.find]
where = ["."]  
include = ["youtube_thumbnail_generator*"]

[tool.setuptools.package-data]
"*" = ["*.txt", "*.md", "*.yml", "*.yaml"]

[tool.black]
line-length = 120
target-version = ['py38', 'py39', 'py310', 'py311']

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "-v --cov=src --cov-report=term-missing"